Hosts rebase

+++ config/tasks.py
@@ -7,9 +7,10 @@ import json
 def build_config():
     try:
         with open("config/constants.py", mode='w') as file:
-            file.truncate()
             exclude_keys = ['decline_reasons']
             configs = Config.objects.all().order_by('id')
+            if configs:
+                file.truncate()
             for config in configs:
                 if config.key not in exclude_keys:
                     file.write('%s = "%s"\n' % (config.key.upper(), config.value))
@@ -18,10 +19,12 @@ def build_config():
 
     try:
         with open("config/versions.json", mode='w') as file:
-            file.truncate()
+
             include_keys = ['android_version', 'ios_version']
             configs = Config.objects.all().order_by('id')
             versions = dict()
+            if configs:
+                file.truncate()
             for config in configs:
                 if config.key in include_keys:
                     versions[config.key] = config.value
